Add Bolded Votes & "Changed" statement
The altered vote in a vote count update is now bolded.

There is now a "Changed" line which also denotes where the altered vote was previously.

When the player is killed following a lynch, the formatting has been improved, the result clarified, and the start of Night is now announced.

I also fixed a typo "apent" --> "apend" in the addplayer() function.