Massive votecount rework
I have removed prev_vote from config.py, I have removed the reset of prev_vote from admin_commands.py.

I have removed the many uses of prev_vote from %vote, %unvote, and %votecount.  Instead, the %vote and %unvote commands now determine the user's previous vote, their current vote, and sends that information along with the voter to the %votecount command.

I added a NOT_VOTING constant to the Config file to make for easy and consistent comparisons in cases where a user is not voting (whether due to unvote or failure to vote)

The %votecount command previously experienced a bug where the wrong user would be lynched and a related bug where the same user would be noted as being voted in the change.  This occurred due to moving a section of code such that the "voted" variable was out of scope.  This has been fixed.

%votecount has now been split into several functions - createVoteCountMessage() && findNotVoting() && returnLynchStatus() && findChangedVote() && playersNeededToLynch().

This improves code readability and modularization & made bug-fixing far easier.