Massive Player Object Rework and Database Skeleton Update
Reworked all commands to use the Player class and signup_list is now a dictionary in the form {member: player} where member is a Discord Member object and player is a Player object.

Added a reset method to the Config class that resets all the static values.  Added a bunch of other minor methods and attributes to make interacting with Config values simpler.

Added "db.py" which contains the DatabaseManger class.  This class will be used to as a go-between for various other functions to interact with the database.

Fixed possible bug involving %newday and votecounts, in which Config.votes was set to a regular dictionary and would likely have caused votes to appear out of order after D1.

I've added a small adjustment to %signuplist so that it displays a number next to the name of the person signing up.

Italicized the changed votes in the "CHANGE: " section of the votecount.

Added the configReset(), votesReset(), and the newDay() methods to Config.  The purpose is to standardize how these things are done.

Added the dev command configReport() to see the values currently in Config at any given time.
Added module/function information comments.

Endgame message is now pinned.

The program now uses the Player.kill() function instead of a kill command.

The votecount function now reports (None) if there are no votes.

The playerlist command now reports (None) if there are no Alive players or Dead players.

Votecount() function now cancels the Day End task.

So many minor fixes and tweaks I've lost count of them all.

Player class rework complete for now.  Oi vey that took forever and a day.